backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic[email]
python-jose[cryptography]
passlib
python-multipart
python-dotenv
redis
google-generativeai
cors
pytest
pytest-asyncio
httpx
aiosqlite
